# openapi.yaml
openapi: 3.1.0
info:
  title: dpl
  description: |
    Deploy server: accepts deploy archives over HTTP, builds container images
    with podman, exports static files from the built image when configured,
    and stores deploy state on disk.
  version: 0.1.0

servers:
  - url: http://127.0.0.1:3000

security:
  - bearerAuth: []

paths:
  /deploy/{name}:
    post:
      summary: Start a deploy
      description: |
        Accepts a tar.gz archive with source code and starts a background
        deploy for the named unit. The request returns immediately with
        `status` set to `building`; the image build continues
        in the background.

        If the archive has a single top-level folder, dpl flattens it
        after extraction.
      operationId: deploy
      parameters:
        - $ref: "#/components/parameters/name"
      requestBody:
        required: true
        content:
          application/gzip:
            schema:
              type: string
              format: binary
              description: tar.gz archive with source code
      responses:
        "202":
          description: Deploy accepted; build started in the background.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildResult"
              example:
                version: 1
                status: building
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Another deploy for this unit is already building.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: unit busy
        "405":
          description: Deploy is not allowed for this unit type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: not allowed

  /deploy/{name}/state:
    get:
      summary: Get deploy state
      description: Reads the current deploy state from disk.
      operationId: getBuildResult
      parameters:
        - $ref: "#/components/parameters/name"
      responses:
        "200":
          description: Current deploy state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildResult"
              examples:
                ready:
                  summary: Successful build
                  value:
                    version: 1
                    status: ready
                building:
                  summary: Build in progress
                  value:
                    version: 1
                    status: building
                failed:
                  summary: Failed build
                  value:
                    version: 1
                    status: failed
                    error: "failed to build image: podman build exited with exit status: 125"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /deploy/{name}/log:
    get:
      summary: Get build log
      description: |
        Returns the build log of the current (latest) deploy version.

        The response includes an `X-Offset` header with the total file size
        in bytes. To incrementally fetch new log content, pass the previously
        received value as the `offset` query parameter — only bytes after that
        offset will be returned.
      operationId: getBuildLog
      parameters:
        - $ref: "#/components/parameters/name"
        - name: offset
          in: query
          required: false
          description: |
            Byte offset to start reading from. When omitted, the entire log
            is returned. Use the `X-Offset` value from a previous response
            to fetch only newly appended content.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Build log content (or empty body if offset >= file size).
          headers:
            X-Offset:
              description: Total size of the log file in bytes. Pass as `offset` to fetch new content.
              schema:
                type: integer
                minimum: 0
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Plain Bearer token from `{config.base}/auth.yaml`.
        Each token has a list of allowed unit names (or `"*"` for all).

  parameters:
    name:
      name: name
      in: path
      required: true
      description: |
        Unit name. Must be non-empty, lowercase ASCII alphanumerics
        and hyphens only, must not start or end with a hyphen, and must
        not contain consecutive hyphens.
      schema:
        type: string
        pattern: ^[a-z0-9]+(-[a-z0-9]+)*$
        examples:
          - myapp

  schemas:
    BuildResult:
      type: object
      required:
        - version
        - status
      properties:
        version:
          type: integer
          minimum: 0
          description: Version number of the latest deploy attempt.
        status:
          $ref: "#/components/schemas/DeployStatus"
        error:
          type: string
          description: Error message from the latest failed deploy. Present only when status is `failed`.

    DeployStatus:
      type: string
      enum:
        - idle
        - building
        - ready
        - failed

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message.

  responses:
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: invalid token
    Forbidden:
      description: Token is valid but not allowed for this unit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: permission denied
    NotFound:
      description: Unit config does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unit not found